+++ Classes/JIMSubscriptionManager.m
@@ -81,12 +81,12 @@
 }
 
 #pragma mark NSTaleView Delegate
-- (int)numberOfRowsInTableView:(NSTableView *)tableView
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)tableView
 {
 	return [requests count];
 }
 
-- (id)tableView:(NSTableView *)tableView objectValueForTableColumn:(NSTableColumn *)tableColumn row:(int)rowIndex
+- (id)tableView:(NSTableView *)tableView objectValueForTableColumn:(NSTableColumn *)tableColumn row:(NSInteger)rowIndex
 {	
 	if([[tableColumn identifier] isEqualToString:@"Name"])
 		return [[(XMPPSubscriptionRequest *)[requests objectAtIndex:rowIndex] fromJID] fullString];
